backend/requirements.txt
Django==6.0.2
djangorestframework==3.15.2
django-filter==24.3
django-cors-headers==4.4.0
django-unfold==0.81.0
python-decouple==3.8
gunicorn==23.0.0
